--- jasmine.json
@@ -1,5 +1,0 @@
-{
-  "random":                       true,
-  "spec_files":                   ["*.test.js"],
-  "stopSpecOnExpectationFailure": true
-}
