+++ src/index.js
@@ -1,5 +1,14 @@
-function scription2dlx(scription) {
-  return scription;
-}
+import getMatches from './getMatches.js';
 
-export default scription2dlx;
+const headerRegExp = /(---.+---)/su;
+
+/**
+ * The top-level script for this library. Converts a Scription-format text to Daffodil format.
+ * @param  {String} scription The text of the Scription file to parse
+ * @return {Object}           Returns a Daffodil Text object
+ */
+export default function scription2dlx(scription) {
+
+  return text;
+
+}
