+++ test/lines/literal.test.js
@@ -1,0 +1,20 @@
+/**
+ * This file applies tests to the literal translation line (`\lit`)
+ */
+
+const convert = require(`../../scription2dlx`);
+
+describe(`literal translation`, () => {
+
+  it(`removes brackets`, () => {
+
+    const text = `
+     \\txn ninakupenda
+     \\lit [I love you]
+     `;
+
+  });
+
+  it(`may be in multiple languages`);
+
+});
