+++ test/lines/translation.test.js
@@ -1,0 +1,11 @@
+/**
+ * This file applies tests to the free translation line (`\tln`)
+ */
+
+const convert = require(`../../scription2dlx`);
+
+describe(`free translation`, () => {
+
+  it(`may be in multiple languages`);
+
+});
