+++ test/text.test.js
@@ -1,0 +1,13 @@
+/**
+ * This test suite applies tests that check the text as a whole
+ */
+
+const convert = require(`../scription2dlx`);
+
+describe(`text`, () => {
+
+  it(`may be blank`, () => {
+    const text = ``;
+  });
+
+});
